requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
google-generativeai==0.3.2
sentence-transformers==2.2.2
inngest==0.2.3
llama-index-core==0.10.11
llama-index-readers-file==0.1.4
python-dotenv==1.0.0
qdrant-client==1.7.0
streamlit==1.28.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pathlib==1.0.1
langgraph==0.0.26